feat: Add iOS Whisper Demo App
- Added `ios-whisper-demo` directory with `Sources` and `scripts`.
- Implemented `AudioRecorder.swift` for 16kHz audio capture with thread-safe `rawSamples` access.
- Implemented `WhisperState.swift` for `whisper.cpp` integration, handling C-API type safety.
- Created SwiftUI `ContentView` and `App` entry point.
- Added `download_models.sh` script to fetch Whisper models.
- Added comprehensive `README.md` with manual setup instructions.
- Configured `.gitignore` to exclude model binaries and Xcode project files.

+++ ios-whisper-demo/Sources/WhisperState.swift
@@ -100,8 +100,8 @@ class WhisperState: ObservableObject {
             // Extract text
             let n_segments = whisper_full_n_segments(ctx)
             var result = ""
-            for i in 0..<n_segments {
-                if let textPtr = whisper_full_get_segment_text(ctx, i) {
+            for i in 0..<Int(n_segments) {
+                if let textPtr = whisper_full_get_segment_text(ctx, Int32(i)) {
                     let text = String(cString: textPtr)
                     result += text
                 }
